+++ examples/servers/rtu.rs
@@ -7,12 +7,12 @@ use lazy_static::lazy_static;
 use std::sync::RwLock;
 
 use rmodbus::{
-    server::{context::ModbusContext, ModbusFrame},
+    server::{context::ModbusContextFull, ModbusFrame},
     ModbusFrameBuf, ModbusProto,
 };
 
 lazy_static! {
-    pub static ref CONTEXT: RwLock<ModbusContext> = RwLock::new(ModbusContext::new());
+    pub static ref CONTEXT: RwLock<ModbusContextFull> = RwLock::new(ModbusContextFull::new());
 }
 
 pub fn rtuserver(unit: u8, port: &str) {